Refactor: move SSTI input to physics
