+++ src/global_legacy_module/4C_global_legacy_module_validconditions.cpp
@@ -26,7 +26,6 @@
 #include "4C_inpar_particle.hpp"
 #include "4C_inpar_s2i.hpp"
 #include "4C_inpar_scatra.hpp"
-#include "4C_inpar_ssti.hpp"
 #include "4C_inpar_sti.hpp"
 #include "4C_inpar_structure.hpp"
 #include "4C_inpar_xfem.hpp"
@@ -34,6 +33,7 @@
 #include "4C_porofluid_pressure_based_elast_scatra_input.hpp"
 #include "4C_red_airways_input.hpp"
 #include "4C_ssi_input.hpp"
+#include "4C_ssti_input.hpp"
 #include "4C_thermo_input.hpp"
 
 FOUR_C_NAMESPACE_OPEN
@@ -831,7 +831,7 @@ std::vector<Core::Conditions::ConditionDefinition> Global::valid_conditions()
   Inpar::Solid::set_valid_conditions(condlist);
   Thermo::set_valid_conditions(condlist);
   SSI::set_valid_conditions(condlist);
-  Inpar::SSTI::set_valid_conditions(condlist);
+  SSTI::set_valid_conditions(condlist);
   Inpar::PARTICLE::set_valid_conditions(condlist);
   Inpar::LevelSet::set_valid_conditions(condlist);
   BeamInteraction::Potential::set_valid_conditions(condlist);
