+++ src/global_legacy_module/4C_global_legacy_module_validparameters.cpp
@@ -46,7 +46,6 @@
 #include "4C_inpar_s2i.hpp"
 #include "4C_inpar_scatra.hpp"
 #include "4C_inpar_searchtree.hpp"
-#include "4C_inpar_ssti.hpp"
 #include "4C_inpar_sti.hpp"
 #include "4C_inpar_structure.hpp"
 #include "4C_inpar_volmortar.hpp"
@@ -68,6 +67,7 @@
 #include "4C_red_airways_input.hpp"
 #include "4C_solver_nonlin_nox_input.hpp"
 #include "4C_ssi_input.hpp"
+#include "4C_ssti_input.hpp"
 #include "4C_structure_new_monitor_dbc_input.hpp"
 #include "4C_thermo_input.hpp"
 #include "4C_tsi_input.hpp"
@@ -269,7 +269,7 @@ std::map<std::string, Core::IO::InputSpec> Global::valid_parameters()
   PoroPressureBased::set_valid_parameters_porofluid_elast(specs);
   EHL::set_valid_parameters(specs);
   SSI::set_valid_parameters(specs);
-  Inpar::SSTI::set_valid_parameters(specs);
+  SSTI::set_valid_parameters(specs);
   ALE::set_valid_parameters(specs);
   Inpar::FSI::set_valid_parameters(specs);
